Q3/flash_sale_system/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
